OAUTH2

+++ project/routers/reviews.py
@@ -1,25 +1,24 @@
-from fastapi import HTTPException, APIRouter
+from fastapi import HTTPException, APIRouter, Depends
 
 from ..schemas import ReviewResponseModel, ReviewRequestModel, ReviewRequestPutModel
 
 from ..database import User, Movie, UserReview
+
+from ..common import get_current_user
 
 from typing import List
 
 router = APIRouter(prefix='/reviews')
 
 @router.post('', response_model=ReviewResponseModel)
-async def create_review(user_review: ReviewRequestModel):
-
-    if (User.select().where(User.id == user_review.user_id).first()) is None:
-        raise HTTPException(404, 'User not found')
+async def create_review(user_review: ReviewRequestModel, user: User = Depends(get_current_user)):
     
     if Movie.select().where(Movie.id == user_review.movie_id).first() is None:
         raise HTTPException(404, 'Movie not found')
 
     
     user_review = UserReview.create(
-        user_id=user_review.user_id,
+        user_id=user.id,
         movie_id=user_review.movie_id,
         review=user_review.review,
         score=user_review.score
@@ -50,11 +49,14 @@ async def get_review(review_id: int):
 
 
 @router.put('/{review_id}', response_model=ReviewResponseModel)
-async def update_review(review_id: int, review_request: ReviewRequestPutModel):
+async def update_review(review_id: int, review_request: ReviewRequestPutModel, user: User = Depends(get_current_user)):
     user_review = UserReview.select().where(UserReview.id == review_id).first()
     
     if user_review is None:
         raise HTTPException(404, 'Review not found')
+    
+    if user_review.user_id != user.id:
+        raise HTTPException(401, 'Unauthorized')
     
     user_review.review = review_request.review
     user_review.score = review_request.score
@@ -65,11 +67,14 @@ async def update_review(review_id: int, review_request: ReviewRequestPutModel):
 
 
 @router.delete('/{review_id}', response_model=ReviewResponseModel)
-async def delete_review(review_id: int):
+async def delete_review(review_id: int, user: User = Depends(get_current_user)):
     user_review = UserReview.select().where(UserReview.id == review_id).first()
     
     if user_review is None:
         raise HTTPException(404, 'Review not found')
+    
+    if user_review.user_id != user.id:
+        raise HTTPException(401, 'Not your review')
     
     user_review.delete_instance()
